workflow-orchestration/pyproject.toml
[project]
name = "workflow-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.0",
    "pgcli>=4.4.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "tqdm>=4.67.3",
]
